♻️ [Refactoring]: [FE] AppViewProvider を src/providers/ 配下にフォルダ化
Provider 関連の配置場所として `src/providers/` を新設し、AppViewProvider / Context /
useAppView / 関連型を `src/providers/AppViewProvider/` 配下に集約した。`useAppView`
フックは Context の唯一の正規消費点として Provider と同居させ、`AppViewContext` は
引き続き private を維持する。

- src/hooks/useAppView/index.tsx を src/providers/AppViewProvider/index.tsx に移動
  （内容は不変。型 + Context + Provider + useAppView をすべて同居）
- src/hooks/useAppView/__tests__/useAppView.interaction.test.tsx を
  src/providers/AppViewProvider/__tests__/AppViewProvider.interaction.test.tsx に
  移動（import 元のみ @/providers/AppViewProvider に切替、12 ケース全件 pass を維持）
- src/App.tsx / src/features/board/components/HeaderBar/index.tsx /
  src/hooks/useAppView/resolveCloseTarget.ts の import を @/providers/AppViewProvider に
  切替
- src/hooks/useAppView/ には resolveCloseTarget.ts と resolveCloseTarget の behavior
  テストのみが残る（移動はスコープ外）
