+++ src/App.tsx
@@ -9,17 +9,17 @@ import { LabelRegistry } from "@/domains/label-registry";
 import { Milestone } from "@/domains/milestone";
 import { countTasksWithParseError } from "@/domains/parse-error";
 import { selectTaskOutcome } from "@/domains/task-selection";
-import {
-  type AppView,
-  AppViewProvider,
-  type UseAppViewResult,
-} from "@/hooks/useAppView";
 import { resolveCloseTarget } from "@/hooks/useAppView/resolveCloseTarget";
 import { useLabels } from "@/hooks/useLabels";
 import { useMilestones } from "@/hooks/useMilestones";
 import { useRecentProjects } from "@/hooks/useRecentProjects";
 import { useToasts } from "@/hooks/useToasts";
 import { type OrphanStrategy, registerToastSink } from "@/lib/tauri";
+import {
+  type AppView,
+  AppViewProvider,
+  type UseAppViewResult,
+} from "@/providers/AppViewProvider";
 import { basenameOf } from "@/utils/path";
 import {
   BoardWorkspace,
